// apps/api/src/prefill/lib/schema.ts
export type FieldType = "string" | "integer" | "number" | "boolean" | "date-time" | "uuid" | "enum" | "object" | "array";

export interface FieldDef {
  type: FieldType;
  enumName?: string;
  nullable?: boolean;
  format?: string;
  objectSchema?: Record<string, FieldDef>;
  arrayItemSchema?: Record<string, FieldDef>;
  required?: string[];
}